# Vectorize _count_nodes_freq for NeighborCooccurrenceEncoder in DyGFormer