## Objective

- Vectorize _count_nodes_freq so that it uses tensors on GPU instead of numpy to avoid overhead from CPU and GPU conversion.

*Related:* https://github.com/tgm-team/tgm/issues/216

## Results

**Efficiency**
Dataset: `tgbl-wiki`
GPU: A40

Model speed up
| Version | Train Epoch Latency on | Speedup |
|:---:|:---:|:---:|
| `DyGFormer` with original `NeighborCooccurrenceEncoder`  | 134.4s | --- |
| `DyGFormer` with our optimized `NeighborCooccurrenceEncoder`  | 79.4s | ~1.7x |

`NeighborCooccurrenceEncoder` speed up
| Version | Latency | Speedup |
|:---:|:---:|:---:|
| Original `NeighborCooccurrenceEncoder`  | 0.0468s | --- |
| Our optimized `NeighborCooccurrenceEncoder`  | 0.0046s | ~10x |

**Corectness**
A simple but rigorous test that compare the output of our optimized implementation of `_count_nodes_freq` and original `_count_nodes_freq` from DyGFormer:

```python
full_data = DGData.from_tgb(args.dataset)
full_data = full_data.discretize(args.time_gran)
full_graph = DGraph(full_data)
num_nodes = full_graph.num_nodes
edge_feats_dim = full_graph.edge_feats_dim

full_dg = DGraph(full_data , device=args.device)

nbr_hook = RecencyNeighborHook(
    num_nbrs=[31],  # Keep 1 slot for seed node itself
    num_nodes=num_nodes,
    edge_feats_dim=edge_feats_dim,
)

hm = HookManager(keys=['global'])
hm.register_shared(DeduplicationHook())
hm.register_shared(nbr_hook)

full_loader= DGDataLoader(full_dg , batch_size=args.bsize, hook_manager=hm)

def _count_optimize(src: torch.Tensor, dst: torch.Tensor):
    cross_mask = src.unsqueeze(dim=1) == dst.unsqueeze(dim=-1)
    src_mask = src.unsqueeze(dim=1) == src.unsqueeze(dim=-1)
    dst_mask = dst.unsqueeze(dim=1) == dst.unsqueeze(dim=-1)

    src_freq = torch.stack([src_mask.sum(1), cross_mask.sum(1)], dim=2).float()
    dst_freq = torch.stack([dst_mask.sum(1), cross_mask.sum(2)], dim=2).float()
    src_freq[src == -1] = 0.0
    dst_freq[dst == -1] = 0.0
    return src_freq, dst_freq


neighbour_module = NeighborCooccurrenceEncoder(args.embed_dim, args.device)
with hm.activate('global'):
    for batch in tqdm(full_loader):
        batch_size = batch.src.shape[0]
        src_neighbour = batch.nbr_nids[0][:batch_size]
        dst_neighbour = batch.nbr_nids[0][batch_size : batch_size * 2]
        start = timeit.default_timer()
        source_freq_tensor_exp, dst_freq_tensor_exp = (
            neighbour_module._count_nodes_freq( #original implementation
                src_neighbour.cpu().numpy(), dst_neighbour.cpu().numpy()
            )
        )
        end = timeit.default_timer()
        old_version.append(end - start)

        start = timeit.default_timer()
        source_freq_tensor, dst_freq_tensor = _count_optimize(
            src_neighbour, dst_neighbour
        )
        end = timeit.default_timer()
        new_version.append(end - start)

        assert torch.equal(source_freq_tensor_exp, source_freq_tensor)
        assert torch.equal(dst_freq_tensor_exp, dst_freq_tensor)
```

All `assert` statements passed